+++ vector/src/main/java/im/vector/fragments/VectorSettingsPreferencesFragment.kt
@@ -1240,7 +1240,8 @@ class VectorSettingsPreferencesFragment : PreferenceFragment(), SharedPreference
         if (resultCode == Activity.RESULT_OK) {
             when (requestCode) {
                 REQUEST_NOTIFICATION_RINGTONE -> {
-                    PreferencesManager.setNotificationRingTone(activity, data?.getParcelableExtra<Parcelable>(RingtoneManager.EXTRA_RINGTONE_PICKED_URI) as Uri)
+                    PreferencesManager.setNotificationRingTone(activity,
+                            data?.getParcelableExtra<Parcelable>(RingtoneManager.EXTRA_RINGTONE_PICKED_URI) as Uri?)
 
                     // test if the selected ring tone can be played
                     if (null == PreferencesManager.getNotificationRingToneName(activity)) {